fixes to correct visual c++ warnings, as well as some compile errors.

+++ include/carve/geom2d.hpp
@@ -132,7 +132,7 @@ namespace carve {
 
     template<typename P2vec>
     bool lineIntersectsTriangle(const P2 &p1, const P2 &p2, const P2vec &tri) {
-      int s[3];
+      double s[3];
       // does tri lie on one side or the other of p1-p2?
       s[0] = orient2d(p1, p2, tri[0]);
       s[1] = orient2d(p1, p2, tri[1]);
